+++ derivatives/models/bonds/index_linked_bond_forward.py
@@ -2,7 +2,7 @@
 
 from math import exp
 
-from .base import DerivativeModel
+from ..base import DerivativeModel
 
 
 class IndexLinkedBondForward(DerivativeModel):
