+++ derivatives/models/bonds/zero_coupon.py
@@ -2,7 +2,7 @@
 
 from math import exp
 
-from .base import DerivativeModel
+from ..base import DerivativeModel
 
 
 class ZeroCoupon(DerivativeModel):
