+++ derivatives/models/futures/synthetic_underlying_forward.py
@@ -2,7 +2,7 @@
 
 from math import exp
 
-from .base import DerivativeModel
+from ..base import DerivativeModel
 
 
 class SyntheticUnderlyingForward(DerivativeModel):
