+++ derivatives/models/options/asian_arithmetic_fix_mm.py
@@ -1,13 +1,17 @@
-"""Arithmetic Asian option pricer for commodity underlyings."""
+"""Arithmetic Asian option model with fixed monitoring dates."""
 
 from math import exp
 
 import numpy as np
 
-from .base import DerivativeModel
+from ..base import DerivativeModel
 
-class CommodityAsianOption(DerivativeModel):
-    """Monte Carlo pricer for arithmetic-average Asian options on commodities."""
+class AsianArithmeticFixMM(DerivativeModel):
+    """Monte Carlo pricer for arithmetic-average Asian options.
+
+    Parameters mirror the standard Black--Scholes setting.  Monitoring of the
+    underlying price is assumed to take place at equally spaced intervals.
+    """
 
     def price(
         self,
@@ -17,16 +21,15 @@ class CommodityAsianOption(DerivativeModel):
         vol: float,
         maturity: float,
         num_obs: int,
-        convenience_yield: float = 0.0,
         is_call: bool = True,
         n_paths: int = 10_000,
     ) -> float:
-        """Return the option value.
+        """Return the value of the Asian option.
 
         Parameters
         ----------
         spot:
-            Current spot price of the commodity.
+            Current price of the underlying asset.
         strike:
             Option strike level.
         rate:
@@ -37,12 +40,10 @@ class CommodityAsianOption(DerivativeModel):
             Time to maturity in years.
         num_obs:
             Number of equally spaced observation dates.
-        convenience_yield:
-            Convenience yield of the commodity.
         is_call:
             ``True`` for a call option, ``False`` for a put.
         n_paths:
-            Number of Monte Carlo paths.
+            Number of Monte Carlo paths to simulate.
 
         Returns
         -------
@@ -51,7 +52,7 @@ class CommodityAsianOption(DerivativeModel):
         """
 
         dt = maturity / num_obs
-        drift = (rate - convenience_yield - 0.5 * vol ** 2) * dt
+        drift = (rate - 0.5 * vol ** 2) * dt
         diffusion = vol * np.sqrt(dt)
 
         payoffs = np.zeros(n_paths)
@@ -61,10 +62,10 @@ class CommodityAsianOption(DerivativeModel):
             for _ in range(num_obs):
                 price_path *= np.exp(drift + diffusion * np.random.normal())
                 running_sum += price_path
-            avg_price = running_sum / num_obs
+            average_price = running_sum / num_obs
             if is_call:
-                payoffs[p] = max(avg_price - strike, 0.0)
+                payoffs[p] = max(average_price - strike, 0.0)
             else:
-                payoffs[p] = max(strike - avg_price, 0.0)
+                payoffs[p] = max(strike - average_price, 0.0)
 
         return float(exp(-rate * maturity) * payoffs.mean())
