+++ derivatives/models/options/commodity_asian_option.py
@@ -1,17 +1,13 @@
-"""Arithmetic Asian option model with fixed monitoring dates."""
+"""Arithmetic Asian option pricer for commodity underlyings."""
 
 from math import exp
 
 import numpy as np
 
-from .base import DerivativeModel
+from ..base import DerivativeModel
 
-class AsianArithmeticFixMM(DerivativeModel):
-    """Monte Carlo pricer for arithmetic-average Asian options.
-
-    Parameters mirror the standard Black--Scholes setting.  Monitoring of the
-    underlying price is assumed to take place at equally spaced intervals.
-    """
+class CommodityAsianOption(DerivativeModel):
+    """Monte Carlo pricer for arithmetic-average Asian options on commodities."""
 
     def price(
         self,
@@ -21,15 +17,16 @@ class AsianArithmeticFixMM(DerivativeModel):
         vol: float,
         maturity: float,
         num_obs: int,
+        convenience_yield: float = 0.0,
         is_call: bool = True,
         n_paths: int = 10_000,
     ) -> float:
-        """Return the value of the Asian option.
+        """Return the option value.
 
         Parameters
         ----------
         spot:
-            Current price of the underlying asset.
+            Current spot price of the commodity.
         strike:
             Option strike level.
         rate:
@@ -40,10 +37,12 @@ class AsianArithmeticFixMM(DerivativeModel):
             Time to maturity in years.
         num_obs:
             Number of equally spaced observation dates.
+        convenience_yield:
+            Convenience yield of the commodity.
         is_call:
             ``True`` for a call option, ``False`` for a put.
         n_paths:
-            Number of Monte Carlo paths to simulate.
+            Number of Monte Carlo paths.
 
         Returns
         -------
@@ -52,7 +51,7 @@ class AsianArithmeticFixMM(DerivativeModel):
         """
 
         dt = maturity / num_obs
-        drift = (rate - 0.5 * vol ** 2) * dt
+        drift = (rate - convenience_yield - 0.5 * vol ** 2) * dt
         diffusion = vol * np.sqrt(dt)
 
         payoffs = np.zeros(n_paths)
@@ -62,10 +61,10 @@ class AsianArithmeticFixMM(DerivativeModel):
             for _ in range(num_obs):
                 price_path *= np.exp(drift + diffusion * np.random.normal())
                 running_sum += price_path
-            average_price = running_sum / num_obs
+            avg_price = running_sum / num_obs
             if is_call:
-                payoffs[p] = max(average_price - strike, 0.0)
+                payoffs[p] = max(avg_price - strike, 0.0)
             else:
-                payoffs[p] = max(strike - average_price, 0.0)
+                payoffs[p] = max(strike - avg_price, 0.0)
 
         return float(exp(-rate * maturity) * payoffs.mean())
