+++ derivatives/models/options/discretised_barrier.py
@@ -4,7 +4,7 @@ from math import exp
 
 import numpy as np
 
-from .base import DerivativeModel
+from ..base import DerivativeModel
 
 class DiscretisedBarrier(DerivativeModel):
     """Monte Carlo pricer for discretely monitored barrier options."""
