+++ derivatives/models/options/udmc_cliquet_model.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from typing import Iterable
 
-from .base import DerivativeModel
+from ..base import DerivativeModel
 
 
 class UDMCCliquetModel(DerivativeModel):
