+++ derivatives/models/swaps/total_return_swap.py
@@ -2,7 +2,7 @@
 
 from math import exp
 
-from .base import DerivativeModel
+from ..base import DerivativeModel
 
 
 class TotalReturnSwap(DerivativeModel):
